Towards more complete analysis slices
Analysis slices are ad-hoc. TODO: Create a more formal notion of the 'scope' of an analytic type and highlight the relevant structural syntax to make this scope a valid term.

Note: maintaining the syntax structure with the current slice type is inefficient. A global/non-incremental structure ids slice is beneficial.